+++ src/api/index.ts
@@ -42,3 +42,4 @@ export {
   createApp,
   runServer,
 } from "@/api/server";
+export * from "@/api/middlewares";